refactor: merge skill_create and skill_delete into a single 'skills' tool
Matches the dispatch-on-action style used by gmail/notion/obsidian.
The tool name is 'skills'; actions are 'create' and 'delete'. Hot-reload
behavior and name validation are unchanged.
